+++ resources/views/pages/records/table.blade.php
@@ -23,12 +23,16 @@
                 <td>
                     @if($record->status==="edited")
 
-                        <a href="{{ route('record.list', $record->uuid) }}" title="List Records" class="mr-2">
-                            <em class="icon ni ni-setting-alt" style="font-size: 20px"></em>
-                        </a>
-                        <a href="#" title="Delete Records" class="mr-2 text-danger" onclick="deleteItem('{{ route('record.pop', $record->uuid) }}')">
-                            <em class="icon ni ni-trash-alt" style="font-size: 20px"></em>
-                        </a>
+                        @if($person->hasAccess('edit_record'))
+                            <a href="{{ route('record.list', $record->uuid) }}" title="List Records" class="mr-2">
+                                <em class="icon ni ni-setting-alt" style="font-size: 20px"></em>
+                            </a>
+                        @endif
+                        @if($person->hasAccess('delete_record'))
+                            <a href="#" title="Delete Records" class="mr-2 text-danger" onclick="deleteItem('{{ route('record.pop', $record->uuid) }}')">
+                                <em class="icon ni ni-trash-alt" style="font-size: 20px"></em>
+                            </a>
+                        @endif
 
                         @if($record->ready)
                             <a href="#" title="Start Process" class="ml-2">
